better behaved implicits with .asVector and .asPartialMap methods instead of automatic conversions; automagically vectorized type-preserving versions of everything in Math.{log,exp,cos,round,...}

+++ src/scalala/ScalalaProfiling.scala
@@ -39,7 +39,7 @@ trait ScalalaProfiling {
       for (i <- 0 until n) yield {
         val start = System.currentTimeMillis;
         func;
-        System.currentTimeMillis - start;
+        (System.currentTimeMillis - start).toDouble;
       }
     );
   }
